% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_column.R
\name{normalize_column}
\alias{normalize_column}
\title{BUILD: Normalize a column}
\usage{
normalize_column(data, column)
}
\arguments{
\item{data}{data.frame that holds the column to be normalized}

\item{column}{the column name to normalize}
}
\description{
Normalizes a column so the numeric values are rescaled from 0 to 1
}
