% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_file.R
\name{open_file}
\alias{open_file}
\title{CLEAN: Unzip the file and load data using keyword to obtain secret key}
\usage{
open_file(filename, keyword, hashing_function, serialize = FALSE)
}
\arguments{
\item{filename}{A character value of the location of the file to open}

\item{keyword}{A character value to apply a cryptographical hash function to}

\item{hashing_function}{Cryptographical hash function to apply to keyword, must be one of md5, sha1, crc32, sha256, sha512, xxhash32, xxhash64 and murmur32.}

\item{serialize}{A logical variable indicating whether the object should be serialized using serialize (in ASCII form). Defaults to FALSE}
}
\description{
CLEAN: Unzip the file and load data using keyword to obtain secret key
}
